[IDLE-000] 불필요한 객체 제거 및 NotificationChannel을 생성하는 방법의 책임을 NotificationService로 이동
